==> 2. Concatenate data from individual images.ipynb <==
#Import stuff 
import pandas as pd
import os
import numpy as np
from tkinter import Tk
import glob
print ("Pandas imported")
print ("os imported")
print ("Numpy imported")
print ("tkinter imported")

##### Display dialog box for user to select the path. 

from tkinter.filedialog import askdirectory
inputDir = askdirectory(title='Select Input Directory') 
print("Selected directory: " + inputDir)  
os.chdir(inputDir)
cwd = os.getcwd()
print("Current working directory: {0}".format(cwd))

##### Optional static input directory here, remove hashes. 

#inputDir = '/Users/abbykroken/Desktop/F5a/070920 n2/dfs'
#print("Selected directory: " + inputDir)  
#os.chdir(inputDir)
#cwd = os.getcwd()
#print("Current working directory: {0}".format(cwd))

#Stack all the files in the directory together in one dataframe.

path = inputDir
all_files = glob.glob(path + "/*.csv")

df = pd.concat((pd.read_csv(f) for f in all_files))

#Save dataframe as a .csv file.

df.to_csv('All_Results.csv', index=False)
print('Your data has been saved as All_Results.csv.')
