==> 3. Analyze 17-frame time lapses.ipynb <==
# Import stuff 
import pandas as pd
import os
import numpy as np
print ("Pandas imported")
print ("os imported")
print ("Numpy imported")

#### CHANGE DIRECTORY ####

inputDir = '/Users/abbykroken/Desktop/sand2/dfs'

print("Selected directory: " + inputDir)  
os.chdir(inputDir)
cwd = os.getcwd()
print("Current working directory: {0}".format(cwd))

# Load results file as and examine it.
df = pd.read_csv('All_Results.csv')
pd.options.display.max_columns = None
df

# Create a new dataframe to populate processed results

selected_columns = df[["Unique ID", "Strain"]]
resultdf = selected_columns.copy()
resultdf.head()

# User defines cutoffs. Determine these by viewing the TIF results file and raw data for GFP max.
# 1.6 used often for PI/Hoechst cutoff.
# 10 works well for GFP area.
# 525 works well for GFP max. 

dead = float(input("PI to Hoechst ratio cutoff: "))
gfpArea = int(input("GFP min area required: "))
gfpMax = int(input("GFP Max required for positive signal: "))


# Make a new dataframe for just the cell death.

columnss = df[["Unique ID","04h-n","05h-n","06h-n","07h-n","08h-n","09h-n","10h-n","11h-n","12h-n","13h-n","14h-n","15h-n","16h-n","17h-n","18h-n","19h-n","20h-n"]]

nucdf = columnss.copy()
nucdf.head(15)

# ensure all nucleus data is numeric and rounded

nucdf['04h-n'] = pd.to_numeric(nucdf['04h-n'],errors='coerce')
nucdf['05h-n'] = pd.to_numeric(nucdf['05h-n'],errors='coerce') 
nucdf['06h-n'] = pd.to_numeric(nucdf['06h-n'],errors='coerce')
nucdf['07h-n'] = pd.to_numeric(nucdf['07h-n'],errors='coerce')
nucdf['08h-n'] = pd.to_numeric(nucdf['08h-n'],errors='coerce')
nucdf['09h-n'] = pd.to_numeric(nucdf['09h-n'],errors='coerce')
nucdf['10h-n'] = pd.to_numeric(nucdf['10h-n'],errors='coerce')
nucdf['11h-n'] = pd.to_numeric(nucdf['11h-n'],errors='coerce')
nucdf['12h-n'] = pd.to_numeric(nucdf['12h-n'],errors='coerce')
nucdf['13h-n'] = pd.to_numeric(nucdf['13h-n'],errors='coerce')
nucdf['14h-n'] = pd.to_numeric(nucdf['14h-n'],errors='coerce')
nucdf['15h-n'] = pd.to_numeric(nucdf['15h-n'],errors='coerce')
nucdf['16h-n'] = pd.to_numeric(nucdf['16h-n'],errors='coerce')
nucdf['17h-n'] = pd.to_numeric(nucdf['17h-n'],errors='coerce')
nucdf['18h-n'] = pd.to_numeric(nucdf['18h-n'],errors='coerce')
nucdf['19h-n'] = pd.to_numeric(nucdf['19h-n'],errors='coerce')
nucdf['20h-n'] = pd.to_numeric(nucdf['20h-n'],errors='coerce')

# Replace cell value with "alive" or the timepoint if it is dead. 

nucdf["04h-n"] = np.where(nucdf["04h-n"] > dead, 4, "alive")
nucdf["05h-n"] = np.where(nucdf["05h-n"] > dead, 5, "alive")
nucdf["06h-n"] = np.where(nucdf["06h-n"] > dead, 6, "alive")
nucdf["07h-n"] = np.where(nucdf["07h-n"] > dead, 7, "alive")
nucdf["08h-n"] = np.where(nucdf["08h-n"] > dead, 8, "alive")
nucdf["09h-n"] = np.where(nucdf["09h-n"] > dead, 9, "alive")
nucdf["10h-n"] = np.where(nucdf["10h-n"] > dead, 10, "alive")
nucdf["11h-n"] = np.where(nucdf["11h-n"] > dead, 11, "alive")
nucdf["12h-n"] = np.where(nucdf["12h-n"] > dead, 12, "alive")
nucdf["13h-n"] = np.where(nucdf["13h-n"] > dead, 13, "alive")
nucdf["14h-n"] = np.where(nucdf["14h-n"] > dead, 14, "alive")
nucdf["15h-n"] = np.where(nucdf["15h-n"] > dead, 15, "alive")
nucdf["16h-n"] = np.where(nucdf["16h-n"] > dead, 16, "alive")
nucdf["17h-n"] = np.where(nucdf["17h-n"] > dead, 17, "alive")
nucdf["18h-n"] = np.where(nucdf["18h-n"] > dead, 18, "alive")
nucdf["19h-n"] = np.where(nucdf["19h-n"] > dead, 19, "alive")
nucdf["20h-n"] = np.where(nucdf["20h-n"] > dead, 20, "alive")


nucdf.head(5)

# Turn "alive" cells into NaNs. 

nucdf['04h-n'] = pd.to_numeric(nucdf['04h-n'],errors='coerce')
nucdf['05h-n'] = pd.to_numeric(nucdf['05h-n'],errors='coerce') 
nucdf['06h-n'] = pd.to_numeric(nucdf['06h-n'],errors='coerce')
nucdf['07h-n'] = pd.to_numeric(nucdf['07h-n'],errors='coerce')
nucdf['08h-n'] = pd.to_numeric(nucdf['08h-n'],errors='coerce')
nucdf['09h-n'] = pd.to_numeric(nucdf['09h-n'],errors='coerce')
nucdf['10h-n'] = pd.to_numeric(nucdf['10h-n'],errors='coerce')
nucdf['11h-n'] = pd.to_numeric(nucdf['11h-n'],errors='coerce')
nucdf['12h-n'] = pd.to_numeric(nucdf['12h-n'],errors='coerce')
nucdf['13h-n'] = pd.to_numeric(nucdf['13h-n'],errors='coerce')
nucdf['14h-n'] = pd.to_numeric(nucdf['14h-n'],errors='coerce')
nucdf['15h-n'] = pd.to_numeric(nucdf['15h-n'],errors='coerce')
nucdf['16h-n'] = pd.to_numeric(nucdf['16h-n'],errors='coerce')
nucdf['17h-n'] = pd.to_numeric(nucdf['17h-n'],errors='coerce')
nucdf['18h-n'] = pd.to_numeric(nucdf['18h-n'],errors='coerce')
nucdf['19h-n'] = pd.to_numeric(nucdf['19h-n'],errors='coerce')
nucdf['20h-n'] = pd.to_numeric(nucdf['20h-n'],errors='coerce')

# check to make sure the values were replaced as expected.
pd.options.display.max_columns = None
nucdf.head(15)

# Place minimum value in each row into results table. 
# This is the first timepoint where the cell was reported dead.
# This strategy avoids accidently counting cells where the PI signal drops back to "alive" status.

resultdf['Survival Time'] = nucdf.min(axis=1, skipna=None, level=None, numeric_only=None)
resultdf.head(15)


# Create a new dataframe to calculate invasion results starting with GFP area.

selected_columns2 = df[["Unique ID","04h-a","05h-a","06h-a","07h-a","08h-a","09h-a","10h-a","11h-a","12h-a","13h-a","14h-a","15h-a","16h-a","17h-a","18h-a","19h-a","20h-a"]]
areadf = selected_columns2.copy()
areadf.head(15)

# Ensure all data is numeric and rounded.

areadf['04h-a'] = areadf['04h-a'].astype(float).round(3)
areadf['05h-a'] = areadf['05h-a'].astype(float).round(3)
areadf['06h-a'] = areadf['06h-a'].astype(float).round(3)
areadf['07h-a'] = areadf['07h-a'].astype(float).round(3)
areadf['08h-a'] = areadf['08h-a'].astype(float).round(3)
areadf['09h-a'] = areadf['09h-a'].astype(float).round(3)
areadf['10h-a'] = areadf['10h-a'].astype(float).round(3)
areadf['11h-a'] = areadf['11h-a'].astype(float).round(3)
areadf['12h-a'] = areadf['12h-a'].astype(float).round(3)
areadf['13h-a'] = areadf['13h-a'].astype(float).round(3)
areadf['14h-a'] = areadf['14h-a'].astype(float).round(3)
areadf['15h-a'] = areadf['15h-a'].astype(float).round(3)
areadf['16h-a'] = areadf['16h-a'].astype(float).round(3)
areadf['17h-a'] = areadf['17h-a'].astype(float).round(3)
areadf['18h-a'] = areadf['18h-a'].astype(float).round(3)
areadf['19h-a'] = areadf['19h-a'].astype(float).round(3)
areadf['20h-a'] = areadf['20h-a'].astype(float).round(3)

areadf['04h-a'] = pd.to_numeric(areadf['04h-a'],errors='coerce')
areadf['05h-a'] = pd.to_numeric(areadf['05h-a'],errors='coerce') 
areadf['06h-a'] = pd.to_numeric(areadf['06h-a'],errors='coerce')
areadf['07h-a'] = pd.to_numeric(areadf['07h-a'],errors='coerce')
areadf['08h-a'] = pd.to_numeric(areadf['08h-a'],errors='coerce')
areadf['09h-a'] = pd.to_numeric(areadf['09h-a'],errors='coerce')
areadf['10h-a'] = pd.to_numeric(areadf['10h-a'],errors='coerce')
areadf['11h-a'] = pd.to_numeric(areadf['11h-a'],errors='coerce')
areadf['12h-a'] = pd.to_numeric(areadf['12h-a'],errors='coerce')
areadf['13h-a'] = pd.to_numeric(areadf['13h-a'],errors='coerce')
areadf['14h-a'] = pd.to_numeric(areadf['14h-a'],errors='coerce')
areadf['15h-a'] = pd.to_numeric(areadf['15h-a'],errors='coerce')
areadf['16h-a'] = pd.to_numeric(areadf['16h-a'],errors='coerce')
areadf['17h-a'] = pd.to_numeric(areadf['17h-a'],errors='coerce')
areadf['18h-a'] = pd.to_numeric(areadf['18h-a'],errors='coerce')
areadf['19h-a'] = pd.to_numeric(areadf['19h-a'],errors='coerce')
areadf['20h-a'] = pd.to_numeric(areadf['20h-a'],errors='coerce')




# Replace GFP values with dead/NaN where the cell is dead so they are not factored in.

#### SPECIAL INSTRUCTIONS ####
# Hour 4 or up to 6 for rhamnose induction may be left alone here to detect GFP+ bacteria in already-dead cells. 
# Hash out timepoints from 04 to 06 to prevent those from being lost.
# This does not impact the calculation of the cell death point.

#areadf["04h-a"] = np.where(df["04h-n"] < dead, areadf["04h-a"], "dead")
areadf["05h-a"] = np.where(df["05h-n"] < dead, areadf["05h-a"], "dead")
areadf["06h-a"] = np.where(df["06h-n"] < dead, areadf["06h-a"], "dead")
areadf["07h-a"] = np.where(df["07h-n"] < dead, areadf["07h-a"], "dead")
areadf["08h-a"] = np.where(df["08h-n"] < dead, areadf["08h-a"], "dead")
areadf["09h-a"] = np.where(df["09h-n"] < dead, areadf["09h-a"], "dead")
areadf["09h-a"] = np.where(df["09h-n"] < dead, areadf["09h-a"], "dead")
areadf["10h-a"] = np.where(df["10h-n"] < dead, areadf["10h-a"], "dead")
areadf["11h-a"] = np.where(df["11h-n"] < dead, areadf["11h-a"], "dead")
areadf["12h-a"] = np.where(df["12h-n"] < dead, areadf["12h-a"], "dead")
areadf["13h-a"] = np.where(df["13h-n"] < dead, areadf["13h-a"], "dead")
areadf["14h-a"] = np.where(df["14h-n"] < dead, areadf["14h-a"], "dead")
areadf["15h-a"] = np.where(df["15h-n"] < dead, areadf["15h-a"], "dead")
areadf["16h-a"] = np.where(df["16h-n"] < dead, areadf["16h-a"], "dead")
areadf["17h-a"] = np.where(df["17h-n"] < dead, areadf["17h-a"], "dead")
areadf["18h-a"] = np.where(df["18h-n"] < dead, areadf["18h-a"], "dead")
areadf["19h-a"] = np.where(df["19h-n"] < dead, areadf["19h-a"], "dead")
areadf["20h-a"] = np.where(df["20h-n"] < dead, areadf["20h-a"], "dead")

pd.options.display.max_columns = None
areadf.head(15)

# Replace GFP area values with "bg" for background, where the GFP max does not reach the user-entered cutoff. 

areadf["04h-a"] = np.where(df["04h-m"] > gfpMax, areadf["04h-a"], "bg")
areadf["05h-a"] = np.where(df["05h-m"] > gfpMax, areadf["05h-a"], "bg")
areadf["06h-a"] = np.where(df["06h-m"] > gfpMax, areadf["06h-a"], "bg")
areadf["07h-a"] = np.where(df["07h-m"] > gfpMax, areadf["07h-a"], "bg")
areadf["08h-a"] = np.where(df["08h-m"] > gfpMax, areadf["08h-a"], "bg")
areadf["09h-a"] = np.where(df["09h-m"] > gfpMax, areadf["09h-a"], "bg")
areadf["09h-a"] = np.where(df["09h-m"] > gfpMax, areadf["09h-a"], "bg")
areadf["10h-a"] = np.where(df["10h-m"] > gfpMax, areadf["10h-a"], "bg")
areadf["11h-a"] = np.where(df["11h-m"] > gfpMax, areadf["11h-a"], "bg")
areadf["12h-a"] = np.where(df["12h-m"] > gfpMax, areadf["12h-a"], "bg")
areadf["13h-a"] = np.where(df["13h-m"] > gfpMax, areadf["13h-a"], "bg")
areadf["14h-a"] = np.where(df["14h-m"] > gfpMax, areadf["14h-a"], "bg")
areadf["15h-a"] = np.where(df["15h-m"] > gfpMax, areadf["15h-a"], "bg")
areadf["16h-a"] = np.where(df["16h-m"] > gfpMax, areadf["16h-a"], "bg")
areadf["17h-a"] = np.where(df["17h-m"] > gfpMax, areadf["17h-a"], "bg")
areadf["18h-a"] = np.where(df["18h-m"] > gfpMax, areadf["18h-a"], "bg")
areadf["19h-a"] = np.where(df["19h-m"] > gfpMax, areadf["19h-a"], "bg")
areadf["20h-a"] = np.where(df["20h-m"] > gfpMax, areadf["20h-a"], "bg")

pd.options.display.max_columns = None
areadf.head(15)                     

# Make it all numbers again and NaNs again.

areadf['04h-a'] = pd.to_numeric(areadf['04h-a'],errors='coerce')
areadf['05h-a'] = pd.to_numeric(areadf['05h-a'],errors='coerce') 
areadf['06h-a'] = pd.to_numeric(areadf['06h-a'],errors='coerce')
areadf['07h-a'] = pd.to_numeric(areadf['07h-a'],errors='coerce')
areadf['08h-a'] = pd.to_numeric(areadf['08h-a'],errors='coerce')
areadf['09h-a'] = pd.to_numeric(areadf['09h-a'],errors='coerce')
areadf['10h-a'] = pd.to_numeric(areadf['10h-a'],errors='coerce')
areadf['11h-a'] = pd.to_numeric(areadf['11h-a'],errors='coerce')
areadf['12h-a'] = pd.to_numeric(areadf['12h-a'],errors='coerce')
areadf['13h-a'] = pd.to_numeric(areadf['13h-a'],errors='coerce')
areadf['14h-a'] = pd.to_numeric(areadf['14h-a'],errors='coerce')
areadf['15h-a'] = pd.to_numeric(areadf['15h-a'],errors='coerce')
areadf['16h-a'] = pd.to_numeric(areadf['16h-a'],errors='coerce')
areadf['17h-a'] = pd.to_numeric(areadf['17h-a'],errors='coerce')
areadf['18h-a'] = pd.to_numeric(areadf['18h-a'],errors='coerce')
areadf['19h-a'] = pd.to_numeric(areadf['19h-a'],errors='coerce')
areadf['20h-a'] = pd.to_numeric(areadf['20h-a'],errors='coerce')

pd.options.display.max_columns = None
areadf.head(15)

# Make a new column that calculates the average area over all the timepoints.

areadf['Invasion Ave'] = areadf[['04h-a', '05h-a', '06h-a', '07h-a', '08h-a', '09h-a', '10h-a', '11h-a', '12h-a', '13h-a', '14h-a', '15h-a', '16h-a', '17h-a', '18h-a', '19h-a', '20h-a']].mean(axis=1)
pd.options.display.max_columns = None
areadf.head(15)

# Categorize each cell as "invaded" or "not invaded" based on the user-entered minimum GFP area.
# Store this in a new column "Invasion Status"

areadf['Invasion Status'] = np.where(areadf["Invasion Ave"] > gfpArea, "Invaded", "Not Invaded")
pd.options.display.max_columns = None
areadf.head(15)

# Put Invasion status on the results table.

resultdf['Invasion Status'] = areadf[['Invasion Status']]
resultdf.head(15)

# Save results as a new .csv file. 

resultdf.to_csv('Processed_Results.csv' , index=False)
print('Your data has been saved as Processed_Results.csv')

# Save the finalized area results for further analysis if desired.

areadf['Strain'] = df['Strain'] 
areadf.to_csv('GFP_Area_Results.csv' , index=False)
print('Your data has been saved as GFP_Area_Results.csv')
