==> 4. Sum GFP accumulation over time.ipynb <==
# Import stuff 
import pandas as pd
import os
print ("Pandas imported")
print ("os imported")

### CHANGE THE DIRECTORY ###

inputDir = '/Users/abbykroken/Desktop/sand2/dfs'
print("Selected directory: " + inputDir)  
os.chdir(inputDir)
cwd = os.getcwd()
print("Current working directory: {0}".format(cwd))

# Load the GFP area results file.

df = pd.read_csv('GFP_Area_Results.csv')
pd.options.display.max_columns = None
df.head()

# List out all the unique strains in the data set to check for any data entry issues. 

NoStrains = df.Strain.unique()
print(df.Strain.unique())

#Count the number of unique strains.

len(NoStrains)

# Groupby adds all the numerical results from each Strain together. 

sumdf = df.groupby(["Strain"]).sum().reset_index()
sumdf

# Save the finalized area results sum.

sumdf.to_csv('GFP_Area_Sums.csv' , index=False)
print('Your data has been saved as GFP_Area_Results.csv')
